requirements.txt
bottle==0.12.18
bottle-mysql==0.3.1
bottlejwt==0.0.14
boddle==0.2.8
mysqlclient==1.4.6
yoyo-migrations==6.1.0
pymysql==0.9.3
requests==2.22.0
